requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
Django==4.2.3
django-compat==1.0.15
django-cors-headers==4.2.0
django-etc==1.4.0
django-hitcount==1.3.5
django4-background-tasks==1.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gevent==23.7.0
greenlet==2.0.2
hiredis==2.2.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.3.1
msgpack==1.0.5
Pillow==10.0.0
prompt-toolkit==3.0.39
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
razorpay==1.3.0
redis==4.6.0
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
zope.event==5.0
zope.interface==6.0
